snapshot-agent: move the GPU-CR sweeper to the utils package
backends/gc.go -> utils/gpu-cr_utils.go (tests likewise), StartGC ->
StartGPUCRSweeper. The sweeper is GPU-CR hygiene shared by any GPU-CR-driven
backend, not backend logic itself, so it lives with the other node-level
helpers.
