+++ pkg/snapshot-agent/utils/gpu-cr_utils_test.go
@@ -1,4 +1,4 @@
-package backends
+package utils
 
 import (
 	"os"
